+++ src/peer/component/watchers.rs
@@ -163,7 +163,6 @@ impl Component {
             &track_id,
             new_sender.receivers().into_iter().collect(),
             state.connection_mode,
-            peer.peer.connection_state(),
         ));
         let sender = sender::Sender::new(
             &new_sender,
@@ -204,7 +203,6 @@ impl Component {
             &track_id,
             HashSet::from([rcvr_state.sender_id().clone()]),
             state.connection_mode,
-            peer.peer.connection_state(),
         );
         let receiver = receiver::Receiver::new(
             &rcvr_state,
@@ -238,7 +236,6 @@ impl Component {
             &val.0,
             val.1,
             state.connection_mode,
-            peer.peer.connection_state(),
         ));
 
         state.maybe_update_connections.set(None);